server/requirements.txt
pymongo
cryptography
python-dotenv
shortuuid
pyjwt
requests